watermarkbackend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
stegano
